+++ docs/meta/history.md
@@ -117,6 +117,7 @@ This is the first release candidate for Hatch v1, a complete rewrite.
 ***Added:***
 
 - Always include the default build script location for source distributions
+- Automatically remove fields from `project.dynamic` that were added by metadata hooks
 
 ### [0.15.0](https://github.com/ofek/hatch/releases/tag/hatchling-v0.15.0) - 2022-02-23 ### {: #hatchling-v0.15.0 }